Prevent removal of system pages by normal users, fix #6

+++ controllers/PageController.php
@@ -382,7 +382,11 @@ class PageController extends Controller
     {
         $model = $this->findModel($id);
         
-        try {
+        try {                    
+            // Only Superadmin can delete system pages
+            if ($model->type == Page::TYPE_SYSTEM && !Yii::$app->user->can('Superadmin'))
+                throw new \yii\base\Exception(Yii::t('app', 'You do not have the right permissions to delete this item'));
+        
             $transaction = Yii::$app->db->beginTransaction();
             $model->delete();
             $transaction->commit();
